Daemon thread support in C
Native threads that are not spawned using java.lang.Thread are assumed to be daemon threads.  The program will terminate immediately when the last non-daemon thread terminates.

The finalizer thread is now a daemon thread & since the Boehm garbage collector thread is native, it is now a daemon thread as desired.

git-svn-id: http://xmlvm.svn.sourceforge.net/svnroot/xmlvm/trunk/xmlvm@1918 0874c2aa-48f6-4a0f-8a6b-e32f4465a257